+++ tests/testthat/test-05-zvariant.R
@@ -170,12 +170,16 @@ test_that("Assigning and fetching group names and identities works", {
   # Variant with no groups
   no.grps.bat.file <- file.path(.options$setup.dir, 
                                 "03_boundary_length_penalty.bat")
-  no.grps.results.variant <- new("Zvariant", bat.file=no.grps.bat.file)
-  
+  no.grps.results.variant <- new("Zvariant", bat.file = no.grps.bat.file)
   
   # Variant doesn't have groups, so there should be no group names either
   expect_true(is.na(groupnames(no.grps.results.variant)),
               "Test variant group names not NA although they haven't been set")
+  
+  # Test that setting groups when there were initially none works
+  groups(no.grps.results.variant) <- correct.grp.codes
+  expect_identical(groups(no.grps.results.variant), correct.grp.codes, 
+                   paste("Test variant group codes incorrect"))
   
   # Test assigning correct group names and codes
   correct.grp.names <- c("mammals", "owls")