+++ NEWS.md
@@ -1,5 +1,11 @@
 ## CHANGES IN VERSION 0.4.2 (2015-XX-XX)
 
+### BUG FIXES
+
++ Assigning groups to a `Zvariant` object when no groups were used initially
+failed before (see [issue #45](https://github.com/cbig/zonator/issues/45).), 
+fixed now.
+
 ### OTHER
 
 + `create_zproject()` no longer uses "do_" prefix when generating bat files.